Temporary fix for #214
A proper fix is required here, probably involving putting the contact details in as part of the audition itself rather than as another pseudo-audition entity, but this is quick fix for now. This is a very partial revert of d31b40ecbb5f8f86bc95a44af1cdd97fd66a634e.