Updates CUADC mailing list info, partially reverts 20f4b68
This commit partially reverts 20f4b68 in that it removes the explicit day of week that the lists are sent on as this tends to vary from each new committee to committee. It also adds contact information from the SM Rep and Techies' Rep who are both now responsible along with the TD for the Techie's List.