Update copyright header
ACTS has been disbanded for many years now and in any event was never an
entity with legal personality. We do not ask code authors, reviewers or
committers to assign copyright thus individuals maintain their own
personal copyright to particular parts of the Camdram project and its
code; the new wording introduced in this commit makes that clear.

Signed-off-by: Charlie Jonas <charlie@charliejonas.co.uk>